+++ tests/test_dewpoint.py
@@ -1,9 +1,9 @@
 import numpy as np
-from utils import ListSource
-from utils import convert_to_ekd_fieldlist
 
 from anemoi.transform.filters import filter_registry
 from anemoi.transform.sources import source_registry
+from anemoi.transform.testing import ListSource
+from anemoi.transform.testing import convert_to_ekd_fieldlist
 
 prototype = {
     "latitudes": [10.0, 0.0, -10.0],
